// src/lib/api/__generated/openapi.ts/zod.gen.ts
export const zTemplateRequest = z.object({
    name: z.string(),
    instanceName: z.string().optional(),
    bossKey: z.string(),
    bossName: z.string(),
    generalNotes: z.string().optional(),
    tankTargetLabels: z.array(z.string()).optional(),
    healGroupLabels: z.array(z.string()).optional(),
    interruptLabels: z.array(z.string()).optional()
});
// ...
export const zCreateEncounterTemplateBody = zTemplateRequest;
// ...
export const zUpdateEncounterTemplateBody = zTemplateRequest;